% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyhwt-package.R
\docType{package}
\name{shinyhwt-package}
\alias{shinyhwt}
\alias{shinyhwt-package}
\title{shinyhwt: Highlighting for Text Inputs in Shiny and htmlwidgets}
\description{
Enables highlighting of text inside text inputs to improve the user experience. Highlights texts based on strings, regular expressions or R objects. Wraps the highlighting-within-textinputs javascript library for shiny.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.maxheld.de/shinyjwt}
  \item \url{https://github.com/maxheld83/shinyhwt}
  \item Report bugs at \url{https://github.com/maxheld83/shinyhwt/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Held \email{info@maxheld.de} (\href{https://orcid.org/0000-0002-4703-5388}{ORCID})

}
\keyword{internal}
